+++ .gitignore
@@ -1,0 +1,2 @@
+testdoc.py
+app2.py
